Python/iot_systemintegration_api_server/requirements.txt
flask-swagger-ui==4.11.1
flask==3.0.0
paho-mqtt==1.6.1
Pillow==10.1.0
requests==2.31.0
urllib3==2.1.0
